add changes

+++ app/Http/Services/V1/Agency/CustomerService.php
@@ -5,6 +5,7 @@ namespace App\Http\Services\V1\Agency;
 
 use App\Exceptions\V1\ModelException;
 use App\Exceptions\V1\FailureException;
+use App\Exceptions\V1\RequestValidationException;
 use App\Http\Services\V1\Customer\CustomerService as AgencyCustomerService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -33,7 +34,7 @@ class CustomerService
             throw FailureException::serverError();
         }
 
-        self::assignUserRole($request,$user);
+        self::assignUserRole($request, $user);
 
         (new UserVerificationService())->generateVerificationCode($user);
 
@@ -55,7 +56,7 @@ class CustomerService
             if (is_array($fullName)) {
                 $users->whereRaw("CONCAT(TRIM(LOWER(first_name)) , ' ' ,TRIM(LOWER(last_name))) in ('" . join("', '", $fullName) . "')");
             } else {
-                $users->whereRaw("CONCAT(LOWER(first_name) , ' ' ,LOWER(last_name)) like ? ", '%'.$fullName.'%');
+                $users->whereRaw("CONCAT(LOWER(first_name) , ' ' ,LOWER(last_name)) like ? ", '%' . $fullName . '%');
             }
         }
 
@@ -65,7 +66,7 @@ class CustomerService
             if (is_array($fname)) {
                 $users->whereRaw("TRIM(LOWER(first_name)) in  ('" . join("', '", $fname) . "')");
             } else {
-                $users->whereRaw('TRIM(LOWER(first_name)) like ?', '%'.$fname.'%');
+                $users->whereRaw('TRIM(LOWER(first_name)) like ?', '%' . $fname . '%');
             }
         }
 
@@ -75,7 +76,7 @@ class CustomerService
             if (is_array($lname)) {
                 $users->whereRaw("TRIM(LOWER(last_name)) in  ('" . join("', '", $lname) . "')");
             } else {
-                $users->whereRaw('TRIM(LOWER(last_name)) like ?', '%'.$lname.'%');
+                $users->whereRaw('TRIM(LOWER(last_name)) like ?', '%' . $lname . '%');
             }
         }
 
@@ -110,8 +111,8 @@ class CustomerService
             $users->whereDate('created_at', '<=', $to);
         }
 
-        $users->whereHas('agencyCustomer',function ($q){
-            $q->where('agency_id',app('agency')->id);
+        $users->whereHas('agencyCustomer', function ($q) {
+            $q->where('agency_id', app('agency')->id);
         });
 
         return ($request->filled('pagination') && $request->get('pagination') == 'false')
@@ -168,6 +169,9 @@ class CustomerService
 
     public static function destroy(User $user)
     {
+        if (!empty($user->agencyCustomer->serviceRequests)) {
+            throw RequestValidationException::errorMessage('Please delete the relational data first.', 422);
+        }
         $user->delete();
     }
 }
